Padrão REST

+++ src/modules/books/implementations/BooksRepository.ts
@@ -4,35 +4,18 @@ import { prisma } from "../../../database";
 import { ICreateBookDTO } from "../dtos/ICreateBookDTO";
 import { IBooksRepository } from "../repositories/IBooksRepository";
 import { UpdateBook } from "../dtos/IUpdateBookDTO";
+import { AppError } from "../../../errors/AppError";
 
 
 class BooksRepository implements IBooksRepository {
     private repository = prisma.book;
 
     async create({
-        title,
-        author,
-        releaseYear,
-        isbn,
-        bannerUrl,
-        loanRate,
-        pages,
-        synopsis,
-        publisher,
-
+        ...bookData
     }: ICreateBookDTO): Promise<void> {
+
         await this.repository.create({
-            data: {
-                title,
-                author,
-                releaseYear,
-                loanRate,
-                isbn,
-                bannerUrl,
-                pages,
-                synopsis,
-                publisher,
-            },
+            data: bookData,
         });
     }
 
@@ -42,11 +25,14 @@ class BooksRepository implements IBooksRepository {
             where: { title },
         });
 
+        if (!books) throw new AppError("Livro não encontrado", 404);
+
         return books;
     }
 
 
     async list({ page = 0, limit = 10, query = '' }): Promise<ListResponse> {
+
 
         const booksLength = await this.repository.count();
 
@@ -74,7 +60,6 @@ class BooksRepository implements IBooksRepository {
             return { ...book, loanId: loan?.id ?? null }
         }))
 
-
         return {
             data: booksResponse,
             total: booksLength,
@@ -83,31 +68,78 @@ class BooksRepository implements IBooksRepository {
             limit
         };
 
+
     }
 
     //Isso aqui provavelmente tá errado
     async update(id: string, new_book: UpdateBook): Promise<Book> {
+
+
+        const book = await this.repository.findUnique({
+            where: { id },
+        });
+
+        if (!book) throw new AppError("Livro não encontrado", 422);
+
+
         const updateresult = await this.repository.update({
             where: {
                 id
             },
             data: {
-                ...new_book
+                ...new_book,
             }
         });
+
+
         return updateresult;
+
     }
 
     async delete(id: string): Promise<Book> {
+
+
+        const book = await this.repository.findUnique({
+            where: { id },
+        });
+
+        if (!book) throw new AppError("Livro não encontrado", 404);
+
+
+        const loan = await prisma.loan.findFirst({
+            where: {
+                bookId: id,
+                status: true
+            },
+            select: {
+                id: true
+            }
+        })
+
+        if (loan) throw new AppError("Livro emprestado", 400);
+
         const deleteresult = await this.repository.delete({
             where: {
                 id
             }
         });
         return deleteresult;
+
     }
 
     async deleteAll(isbn: string): Promise<void> {
+
+        const books = await this.repository.findMany({
+            where: {
+                isbn
+            },
+            select: {
+                id: true
+            }
+        })
+
+        if (books.length === 0) throw new AppError("Livro não encontrado", 404);
+
         await this.repository.deleteMany({
             where: {
                 isbn
